+++ packages/orq-rc/RELEASES.md
@@ -359,3 +359,13 @@ Based on:
 - [typescript v3.1.0-rc.56] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.56 - packages/orq-rc
+
+## 2025-02-11 08:42:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.488.0 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.57] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.57 - packages/orq-rc